Refactor: centralize sorting logic by moving parseSort method to UtilsService
